+++ app/src/main/java/woowacourse/movie/presentation/ui/main/MainPresenterImpl.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.presentation.ui.main
 
 import woowacourse.movie.domain.repository.MovieRepository
-import woowacourse.movie.presentation.dto.MovieUiModel
+import woowacourse.movie.presentation.uimodel.MovieUiModel
 
 class MainPresenterImpl(private val view: MainContract.View, private val movieRepository: MovieRepository) : MainContract.Presenter {
     private var movies: List<MovieUiModel>? = null
